node is configured and working with server

+++ BackupDaemon.py
@@ -12,6 +12,7 @@ from osBackup import performBackup
 
 STATE_BUSY = 1
 STATE_FREE = 0
+CURRENT_STATE = 1
 BACKUP_PROCESS = None
 IO = None
 HOST_NAME = None
@@ -21,9 +22,12 @@ SERVER_PORT = 3000
 
 
 def backup_now():
-    while True:
+    i = 0
+    while i < 3:
         print('Performing a backup...')
-        sleep(3)
+        sleep(6)
+        i += 1
+    emit_new_state(STATE_FREE)
 
 
 def load_config():
@@ -42,57 +46,73 @@ def connect_to_sio():
     print('Connecting...')
     global IO
     IO = SocketIO(SERVER_ADDRESS, SERVER_PORT, LoggingNamespace)
-    IO.emit('join', {'hostname': HOST_NAME})
 
 
 def verify_instance_key():
     # If we do not have an instance key from the server, get one
     if not INSTANCE_KEY:
         IO.on('key_response', on_key_response)
-        IO.emit('key_request')
+        IO.emit('key_request', {'hostname': HOST_NAME})
         IO.wait(seconds=5)
         # Reload the configuration
         print('Reloading configuration...')
         load_config()
         # If the instance key is not available at this point, abort
         if not INSTANCE_KEY: exit(1)
+    IO.emit('join', {'hostname': HOST_NAME, 'instance_key': INSTANCE_KEY})
 
 
 def emit_new_state(state):
     print('Emitting new state...')
+    global CURRENT_STATE
+    CURRENT_STATE = state
     IO.emit('state', {
         'hostname': HOST_NAME,
         'instance_key': INSTANCE_KEY,
-        'state': state
+        'busy': CURRENT_STATE
     })
 
 
 def on_key_response(*args):
     for arg in args:
         # Receive the instance key
-        key = json.loads(arg)['key']
+        key = json.loads(arg)['instance_key']
         print('New instance key received: ' + key)
         conf = readConfig()
         conf['general']['instance_key'] = key
         saveConfig(conf)
 
 
+def on_state_response(*args):
+    for arg in args:
+        success = 'True' if json.loads(arg)['success'] else 'False'
+        print('State notification: ' + success)
+    print('Waiting for direction...')
+
+
 def on_new_backup_task(*args):
-    # Notify the server that you are busy
-    emit_new_state(STATE_BUSY)
-    # Build a new process for the backup job
-    BACKUP_PROCESS = Process(target=backup_now)
-    # Start the job
-    BACKUP_PROCESS.start()
-    # Join it to the pool
-    BACKUP_PROCESS.join()
+    if CURRENT_STATE == STATE_FREE:
+        print('Starting new backup task...')
+        # Notify the server that you are busy
+        emit_new_state(STATE_BUSY)
+        # Build a new process for the backup job
+        BACKUP_PROCESS = Process(target=backup_now)
+        # Start the job
+        BACKUP_PROCESS.start()
+        # Join it to the pool
+        BACKUP_PROCESS.join()
+        return True
+    else:
+        return False
 
 
 def on_end_backup_task(*args):
-    # Terminate the backup task
-    BACKUP_PROCESS.terminate()
-    # Notify the server that you are free
-    emit_new_state(STATE_FREE)
+    if CURRENT_STATE == STATE_BUSY:
+        # Terminate the backup task
+        BACKUP_PROCESS.terminate()
+        # Notify the server that you are free
+        emit_new_state(STATE_FREE)
+    return True
 
 
 def main():
@@ -104,6 +124,8 @@ def main():
     connect_to_sio()
     # Verify that you have a good instance key
     verify_instance_key()
+    # Interpret state response
+    IO.on('state_response', on_state_response)
     # Listen for backup jobs
     IO.on('new_backup_task', on_new_backup_task)
     # Listen for cancel on job
@@ -111,7 +133,6 @@ def main():
     # Notify the server that you are ready
     emit_new_state(STATE_FREE)
     # Hang out
-    print('Waiting for direction...')
     IO.wait()
